// enumeratewebapp/src/model/REC01Info.ts
export interface REC01Info {
    AH_CODE: string;
    A01: string;
    A02_1: string;
    A02_2: string;
    A02_3: string;
    A02_4: string;
    A03: string;
    A04: string;
    A04T: string;
    A05: string;
    A05_1: string;
    A05_2: string;
    A05_3: string;
    A05_4: string;
    A05_5: string;
    A06: string;
    A06_SUB: string;
    A06AMP_SUB: string;
    A06TAM_SUB: string;
    A06VIL_SUB: string;
    A07: string;
    A07A: string;
    A07B: string;
    A07C: string;
    A08: string;
    A08A: string;
    A08B: string;
    A08C: string;
    A09: string;
    A09A: string;
    A09B: string;
    A09C: string;
    A10: string;
    A10A: string;
    A10B: string;
    A10C: string;
    A11: string;
    A11A: string;
    A11B: string;
    A11C: string;
    A12: string;
    A12A: string;
    A12B: string;
    A12C: string;
    A13: string;
    A13A: string;
    A13B: string;
    A13C: string;
    A14: string;
    A14A: string;
    A14B: string;
    A14C: string;
    A15: string;
    A15A: string;
    A15B: string;
    A15C: string;
    A16: string;
    A16A: string;
    A16B: string;
    A16C: string;
    A17: string;
    A17A: string;
    A17B: string;
    A17C: string;
    A18: string;
    A18A: string;
    A18B: string;
    A18C: string;
    A19: string;
    A19A: string;
    A19B: string;
    A19C: string;
    A20: string;
    A20A: string;
    A20B: string;
    A20C: string;
    A21: string;
    A21A: string;
    A21B: string;
    A21C: string;
    A22: string;
    A22A: string;
    A22B: string;
    A22C: string;
    A23: string;
    A23A: string;
    A23B: string;
    A23C: string;
    A24: string;
    A24A: string;
    A24B: string;
    A24C: string;
    A25: string;
    A25A: string;
    A25B: string;
    A25C: string;
    A06VILN_SUB: string;
  }